requirements.txt
# Video processing dependencies
moviepy>=1.0.3
ffmpeg-python>=0.2.0

# Environment and CLI
python-dotenv>=1.0.0

# Utilities
Pillow>=10.0.0

# API clients for social media uploads
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
requests>=2.31.0

